Rename project to projectassetlocator, introduce owning project object
