+++ src/StudioCore/AssetLocator.cs
@@ -88,14 +88,14 @@ public class AssetLocator
     /// <summary>
     ///     The game interroot where all the game assets are
     /// </summary>
-    public string GameRootDirectory => Locator.ActiveProject.ParentProject.RootDirectory;
+    public string GameRootDirectory => Locator.ActiveProject.ParentProject.AssetLocator.RootDirectory;
 
     /// <summary>
     ///     An optional override mod directory where modded files are stored
     /// </summary>
-    public string GameModDirectory => Locator.ActiveProject.RootDirectory;
+    public string GameModDirectory => Locator.ActiveProject.AssetLocator.RootDirectory;
 
-    public string GetAssetPath(string relpath) => Locator.ActiveProject.GetAssetPath(relpath);
+    public string GetAssetPath(string relpath) => Locator.ActiveProject.AssetLocator.GetAssetPath(relpath);
 
     public void SetFromProjectSettings(ProjectSettings settings, string moddir)
     {
@@ -109,110 +109,110 @@ public class AssetLocator
     ///     it will be in this list.
     /// </summary>
     /// <returns></returns>
-    public List<string> GetFullMapList() => Locator.ActiveProject.GetFullMapList();
+    public List<string> GetFullMapList() => Locator.ActiveProject.AssetLocator.GetFullMapList();
 
-    public AssetDescription GetMapMSB(string mapid, bool writemode = false) => Locator.ActiveProject.GetMapMSB(mapid, writemode);
+    public AssetDescription GetMapMSB(string mapid, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetMapMSB(mapid, writemode);
 
-    public List<AssetDescription> GetMapBTLs(string mapid, bool writemode = false) => Locator.ActiveProject.GetMapBTLs(mapid, writemode);
+    public List<AssetDescription> GetMapBTLs(string mapid, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetMapBTLs(mapid, writemode);
 
-    public AssetDescription GetMapNVA(string mapid, bool writemode = false) => Locator.ActiveProject.GetMapNVA(mapid, writemode);
+    public AssetDescription GetMapNVA(string mapid, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetMapNVA(mapid, writemode);
 
     /// <summary>
     ///     Get folders with msgbnds used in-game
     /// </summary>
     /// <returns>Dictionary with language name and path</returns>
-    public Dictionary<string, string> GetMsgLanguages() => Locator.ActiveProject.GetMsgLanguages();
+    public Dictionary<string, string> GetMsgLanguages() => Locator.ActiveProject.AssetLocator.GetMsgLanguages();
 
     /// <summary>
     ///     Get path of item.msgbnd (english by default)
     /// </summary>
-    public AssetDescription GetItemMsgbnd(string langFolder, bool writemode = false) => Locator.ActiveProject.GetItemMsgbnd(langFolder, writemode);
+    public AssetDescription GetItemMsgbnd(string langFolder, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetItemMsgbnd(langFolder, writemode);
 
     /// <summary>
     ///     Get path of menu.msgbnd (english by default)
     /// </summary>
-    public AssetDescription GetMenuMsgbnd(string langFolder, bool writemode = false) => Locator.ActiveProject.GetMenuMsgbnd(langFolder, writemode);
+    public AssetDescription GetMenuMsgbnd(string langFolder, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetMenuMsgbnd(langFolder, writemode);
 
-    public AssetDescription GetMsgbnd(string msgBndType, string langFolder, bool writemode = false) => Locator.ActiveProject.GetMsgbnd(msgBndType, langFolder, writemode);
+    public AssetDescription GetMsgbnd(string msgBndType, string langFolder, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetMsgbnd(msgBndType, langFolder, writemode);
 
-    public string GetGameIDForDir() => AssetUtils.GetGameIDForDir(Locator.ActiveProject.Type);
+    public string GetGameIDForDir() => AssetUtils.GetGameIDForDir(Locator.ActiveProject.AssetLocator.Type);
 
-    public string GetScriptAssetsCommonDir() => Locator.ActiveProject.GetScriptAssetsCommonDir();
+    public string GetScriptAssetsCommonDir() => Locator.ActiveProject.AssetLocator.GetScriptAssetsCommonDir();
 
-    public string GetScriptAssetsDir() => Locator.ActiveProject.GetScriptAssetsDir();
+    public string GetScriptAssetsDir() => Locator.ActiveProject.AssetLocator.GetScriptAssetsDir();
 
-    public string GetUpgraderAssetsDir() => Locator.ActiveProject.GetUpgraderAssetsDir();
+    public string GetUpgraderAssetsDir() => Locator.ActiveProject.AssetLocator.GetUpgraderAssetsDir();
 
-    public string GetGameOffsetsAssetsDir() => Locator.ActiveProject.GetGameOffsetsAssetsDir();
+    public string GetGameOffsetsAssetsDir() => Locator.ActiveProject.AssetLocator.GetGameOffsetsAssetsDir();
 
-    public string GetParamdefDir() => Locator.ActiveProject.GetParamdefDir();
+    public string GetParamdefDir() => Locator.ActiveProject.AssetLocator.GetParamdefDir();
 
-    public string GetTentativeParamTypePath() => Locator.ActiveProject.GetTentativeParamTypePath();
+    public string GetTentativeParamTypePath() => Locator.ActiveProject.AssetLocator.GetTentativeParamTypePath();
 
-    public string GetParammetaDir() => Locator.ActiveProject.GetParammetaDir();
+    public string GetParammetaDir() => Locator.ActiveProject.AssetLocator.GetParammetaDir();
 
-    public string GetParamNamesDir() => Locator.ActiveProject.GetParamNamesDir();
+    public string GetParamNamesDir() => Locator.ActiveProject.AssetLocator.GetParamNamesDir();
 
-    public string GetStrippedRowNamesPath(string paramName) => Locator.ActiveProject.GetStrippedRowNamesPath(paramName);
+    public string GetStrippedRowNamesPath(string paramName) => Locator.ActiveProject.AssetLocator.GetStrippedRowNamesPath(paramName);
 
-    public PARAMDEF GetParamdefForParam(string paramType) => Locator.ActiveProject.GetParamdefForParam(paramType);
+    public PARAMDEF GetParamdefForParam(string paramType) => Locator.ActiveProject.AssetLocator.GetParamdefForParam(paramType);
 
-    public AssetDescription GetDS2GeneratorParam(string mapid, bool writemode = false) => Locator.ActiveProject.GetDS2GeneratorParam(mapid, writemode);
+    public AssetDescription GetDS2GeneratorParam(string mapid, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetDS2GeneratorParam(mapid, writemode);
 
-    public AssetDescription GetDS2GeneratorLocationParam(string mapid, bool writemode = false) => Locator.ActiveProject.GetDS2GeneratorLocationParam(mapid, writemode);
+    public AssetDescription GetDS2GeneratorLocationParam(string mapid, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetDS2GeneratorLocationParam(mapid, writemode);
 
-    public AssetDescription GetDS2GeneratorRegistParam(string mapid, bool writemode = false) => Locator.ActiveProject.GetDS2GeneratorRegistParam(mapid, writemode);
+    public AssetDescription GetDS2GeneratorRegistParam(string mapid, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetDS2GeneratorRegistParam(mapid, writemode);
 
-    public AssetDescription GetDS2EventParam(string mapid, bool writemode = false) => Locator.ActiveProject.GetDS2EventParam(mapid, writemode);
+    public AssetDescription GetDS2EventParam(string mapid, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetDS2EventParam(mapid, writemode);
 
-    public AssetDescription GetDS2EventLocationParam(string mapid, bool writemode = false) => Locator.ActiveProject.GetDS2EventLocationParam(mapid, writemode);
+    public AssetDescription GetDS2EventLocationParam(string mapid, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetDS2EventLocationParam(mapid, writemode);
 
-    public AssetDescription GetDS2ObjInstanceParam(string mapid, bool writemode = false) => Locator.ActiveProject.GetDS2ObjInstanceParam(mapid, writemode);
+    public AssetDescription GetDS2ObjInstanceParam(string mapid, bool writemode = false) => Locator.ActiveProject.AssetLocator.GetDS2ObjInstanceParam(mapid, writemode);
 
-    public List<AssetDescription> GetMapModels(string mapid) => Locator.ActiveProject.GetMapModels(mapid);
+    public List<AssetDescription> GetMapModels(string mapid) => Locator.ActiveProject.AssetLocator.GetMapModels(mapid);
 
-    public string MapModelNameToAssetName(string mapid, string modelname) => Locator.ActiveProject.MapModelNameToAssetName(mapid, modelname);
+    public string MapModelNameToAssetName(string mapid, string modelname) => Locator.ActiveProject.AssetLocator.MapModelNameToAssetName(mapid, modelname);
 
     /// <summary>
     ///     Gets the adjusted map ID that contains all the map assets
     /// </summary>
     /// <param name="mapid">The msb map ID to adjust</param>
     /// <returns>The map ID for the purpose of asset storage</returns>
-    public string GetAssetMapID(string mapid) => Locator.ActiveProject.GetAssetMapID(mapid);
+    public string GetAssetMapID(string mapid) => Locator.ActiveProject.AssetLocator.GetAssetMapID(mapid);
 
-    public AssetDescription GetMapModel(string mapid, string model) => Locator.ActiveProject.GetMapModel(mapid, model);
+    public AssetDescription GetMapModel(string mapid, string model) => Locator.ActiveProject.AssetLocator.GetMapModel(mapid, model);
 
-    public AssetDescription GetMapCollisionModel(string mapid, string model, bool hi = true) => Locator.ActiveProject.GetMapCollisionModel(mapid, model, hi);
+    public AssetDescription GetMapCollisionModel(string mapid, string model, bool hi = true) => Locator.ActiveProject.AssetLocator.GetMapCollisionModel(mapid, model, hi);
 
-    public List<AssetDescription> GetMapTextures(string mapid) => Locator.ActiveProject.GetMapTextures(mapid);
+    public List<AssetDescription> GetMapTextures(string mapid) => Locator.ActiveProject.AssetLocator.GetMapTextures(mapid);
 
-    public List<string> GetEnvMapTextureNames(string mapid) => Locator.ActiveProject.GetEnvMapTextureNames(mapid);
+    public List<string> GetEnvMapTextureNames(string mapid) => Locator.ActiveProject.AssetLocator.GetEnvMapTextureNames(mapid);
 
-    public AssetDescription GetChrTextures(string chrid) => Locator.ActiveProject.GetChrTextures(chrid);
+    public AssetDescription GetChrTextures(string chrid) => Locator.ActiveProject.AssetLocator.GetChrTextures(chrid);
 
-    public AssetDescription GetMapNVMModel(string mapid, string model) => Locator.ActiveProject.GetMapNVMModel(mapid, model);
+    public AssetDescription GetMapNVMModel(string mapid, string model) => Locator.ActiveProject.AssetLocator.GetMapNVMModel(mapid, model);
 
-    public AssetDescription GetHavokNavmeshes(string mapid) => Locator.ActiveProject.GetHavokNavmeshes(mapid);
+    public AssetDescription GetHavokNavmeshes(string mapid) => Locator.ActiveProject.AssetLocator.GetHavokNavmeshes(mapid);
 
-    public AssetDescription GetHavokNavmeshModel(string mapid, string model) => Locator.ActiveProject.GetHavokNavmeshModel(mapid, model);
+    public AssetDescription GetHavokNavmeshModel(string mapid, string model) => Locator.ActiveProject.AssetLocator.GetHavokNavmeshModel(mapid, model);
 
-    public List<string> GetChrModels() => Locator.ActiveProject.GetChrModels();
+    public List<string> GetChrModels() => Locator.ActiveProject.AssetLocator.GetChrModels();
 
-    public AssetDescription GetChrModel(string chr) => Locator.ActiveProject.GetChrModel(chr);
+    public AssetDescription GetChrModel(string chr) => Locator.ActiveProject.AssetLocator.GetChrModel(chr);
 
-    public List<string> GetObjModels() => Locator.ActiveProject.GetObjModels();
+    public List<string> GetObjModels() => Locator.ActiveProject.AssetLocator.GetObjModels();
 
-    public AssetDescription GetObjModel(string obj) => Locator.ActiveProject.GetObjModel(obj);
+    public AssetDescription GetObjModel(string obj) => Locator.ActiveProject.AssetLocator.GetObjModel(obj);
 
-    public AssetDescription GetObjTexture(string obj) => Locator.ActiveProject.GetObjTexture(obj);
+    public AssetDescription GetObjTexture(string obj) => Locator.ActiveProject.AssetLocator.GetObjTexture(obj);
 
-    public AssetDescription GetAetTexture(string aetid) => Locator.ActiveProject.GetAetTexture(aetid);
+    public AssetDescription GetAetTexture(string aetid) => Locator.ActiveProject.AssetLocator.GetAetTexture(aetid);
 
-    public List<string> GetPartsModels() => Locator.ActiveProject.GetPartsModels();
+    public List<string> GetPartsModels() => Locator.ActiveProject.AssetLocator.GetPartsModels();
 
-    public AssetDescription GetPartsModel(string part) => Locator.ActiveProject.GetPartsModel(part);
+    public AssetDescription GetPartsModel(string part) => Locator.ActiveProject.AssetLocator.GetPartsModel(part);
 
-    public AssetDescription GetPartTextures(string partsId) => Locator.ActiveProject.GetPartTextures(partsId);
+    public AssetDescription GetPartTextures(string partsId) => Locator.ActiveProject.AssetLocator.GetPartTextures(partsId);
 
     /// <summary>
     ///     Converts a virtual path to an actual filesystem path. Only resolves virtual paths up to the bnd level,
@@ -220,5 +220,5 @@ public class AssetLocator
     /// </summary>
     /// <param name="virtualPath"></param>
     /// <returns></returns>
-    public string VirtualToRealPath(string virtualPath, out string bndpath) => Locator.ActiveProject.VirtualToRealPath(virtualPath, out bndpath);
+    public string VirtualToRealPath(string virtualPath, out string bndpath) => Locator.ActiveProject.AssetLocator.VirtualToRealPath(virtualPath, out bndpath);
 }
